+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Configuration/NodeTypes.Document.Page.yaml
@@ -8,3 +8,19 @@
   childNodes:
     main:
       type: 'Neos.Neos:ContentCollection'
+  properties:
+    image:
+      type: Neos\Media\Domain\Model\ImageInterface
+      ui:
+        label: 'Image'
+        reloadIfChanged: true
+        inspector:
+          group: 'document'
+          editorOptions:
+            features:
+              crop: true
+            crop:
+              aspectRatio:
+                locked:
+                  width: 1
+                  height: 1
