+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Resources/Private/Fusion/Document/Page/Page.fusion
@@ -7,7 +7,14 @@ prototype(Neos.TestSite:Document.Page) < prototype(Neos.Neos:Page) {
         main = Neos.Neos:ContentCollection {
             nodePath = 'main'
         }
+        nav = Neos.TestSite:Navigation
+        imageUri = Neos.Neos:ImageUri {
+            asset = ${node.properties.image}
+            maximumHeight = 100
+        }
         renderer = afx`
+            {props.nav}
+            <img class="test-page-image" src={props.imageUri} @if.image={node.properties.image}/>
             <h1>{props.titleEditable}</h1>
             <main class="main">
                 {props.main}
